Optimize memory usage

+++ components/clockwise_clock_ext/Object.h
@@ -3,8 +3,8 @@
 
 struct Object {
   const unsigned short *_image;
-  int _width;
-  int _height;
+  uint8_t _width;
+  uint8_t _height;
 
   Object(const unsigned short *image, int width, int height) {
     this->_image = image;
